+++ docker-compose.yml
@@ -27,4 +27,9 @@ services:
         - "5438:5432"
       volumes: 
         - ./postgres-data:/var/lib/postgresql/data
+<<<<<<< HEAD
         # copy the sql script to create tables
+=======
+        # copy the sql script to create tables
+
+>>>>>>> redo changes
